refactor: split npm.rs into submodules, extract cargo audit helper (#105)
Split npm.rs (787 lines) into npm/ directory with 5 files:
- mod.rs (67 lines): dispatcher + re-exports
- install.rs (249 lines): npm install parser + tests
- audit.rs (270 lines): npm audit parser + tests
- outdated.rs (165 lines): npm outdated parser + tests
- ls.rs (156 lines): npm ls parser + tests

Extract extract_vuln_detail() helper from cargo.rs try_parse_audit_json,
reducing it from 91 to ~69 lines.

Pure structural refactor — zero behavioral changes. All existing tests pass.
